requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
load-dotenv==0.1.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
